Fix gang kills resetting to 0 (#28)
I believe the issue here was that there's harmless duplicates in the sql database and when the RowCount > 1, gang kills weren't even being set. This should fix this issue, but we should also find out why we have duplicate entries in the db. Perhaps setting gang kills id column be unique would help...