Disallow saving of players not completely loaded
We defined `g_bLoaded` as true when the first callback fires, but we should really only consider them loaded after the final (ctkills) callback fires. This *may* prevent the error where CT Kills was getting reset. I believe if a client connected and partially loaded in, but left at the right moment, then they'd be able to save a `0` in for CT Kills.